UI: Refactor Telemetry page
